Move pipeline test to folder
* Add inits to make the discoverable
* Move test code
